# Improve Enumerable.Skip and Enumerable.Take performance