--- a/src/libraries/System.Linq/src/System/Linq/SkipTake.SpeedOpt.cs
+++ b/src/libraries/System.Linq/src/System/Linq/SkipTake.SpeedOpt.cs
@@ -3,7 +3,6 @@
 
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Runtime.InteropServices;
 
 namespace System.Linq
 {
@@ -149,14 +148,9 @@ namespace System.Linq
 
             private static void Fill(IList<TSource> source, Span<TSource> destination, int sourceIndex)
             {
-                if (source is TSource[] sourceArray)
+                if (source.TryGetSpan(out ReadOnlySpan<TSource> sourceSpan))
                 {
-                    sourceArray.AsSpan(sourceIndex, destination.Length).CopyTo(destination);
-                    return;
-                }
-                else if (source is List<TSource> sourceList)
-                {
-                    CollectionsMarshal.AsSpan(sourceList).Slice(sourceIndex, destination.Length).CopyTo(destination);
+                    sourceSpan.Slice(sourceIndex, destination.Length).CopyTo(destination);
                     return;
                 }