@eiriktsarpalis 2025-02-11
What does the performance look like for `IList` sources that are neither `List<T>` or `T[]`?