@EgorBo 2025-02-11
@EgorBot -amd -arm
```cs
using System.Collections.Immutable;
using BenchmarkDotNet.Attributes;

[MemoryDiagnoser]
public class Benchmark
{
    private readonly IList<string> _list = Enumerable.Range(0, 1000)
        .Select(i => i.ToString()).ToImmutableArray();

    [Benchmark]
    public List<string> SkipAndTake() =>
        _list.Skip(200).Take(200).ToList();
}
```