@EgorBo 2025-02-11
I think with help of PGO, these casts are cheap (at least, in monomorphic microbenchmarks). Presumably the overhead is more visible on smaller input and when the input is diverse (so PGO can't help).