@eiriktsarpalis 2025-02-11
> Presumably the overhead is more visible on smaller inputs and when the input is diverse (so PGO can't help).

Perhaps using one benchmark class to run three against three different input types could help here?