@miyaji255 2025-02-11
Even with small inputs there was not much degradation in performance

```
|                            Method | SkipCount | TakeCount |         Mean |      Error |     StdDev |       Median |   Gen0 |   Gen1 | Allocated |
|---------------------------------- |---------- |---------- |-------------:|-----------:|-----------:|-------------:|-------:|-------:|----------:|
|    ReadOnlyCollection_SkipAndTake |         2 |         5 |     57.05 ns |   0.479 ns |   0.400 ns |     57.08 ns | 0.0147 |      - |     192 B |
| ReadOnlyCollection_SkipAndTake_PR |         2 |         5 |     59.40 ns |   1.232 ns |   1.954 ns |     58.57 ns | 0.0147 |      - |     192 B |
|                  List_SkipAndTake |         2 |         5 |     48.11 ns |   0.229 ns |   0.179 ns |     48.12 ns | 0.0147 |      - |     192 B |
|               List_SkipAndTake_PR |         2 |         5 |     50.09 ns |   0.496 ns |   0.440 ns |     49.96 ns | 0.0147 |      - |     192 B |
|                     List_GetRange |         2 |         5 |     11.45 ns |   0.150 ns |   0.133 ns |     11.46 ns | 0.0073 |      - |      96 B |
|    ReadOnlyCollection_SkipAndTake |        20 |        50 |    225.20 ns |   1.150 ns |   1.020 ns |    225.09 ns | 0.0422 |      - |     552 B |
| ReadOnlyCollection_SkipAndTake_PR |        20 |        50 |    227.43 ns |   1.754 ns |   1.465 ns |    227.10 ns | 0.0422 |      - |     552 B |
|                  List_SkipAndTake |        20 |        50 |    169.29 ns |   1.191 ns |   1.055 ns |    169.08 ns | 0.0422 |      - |     552 B |
|               List_SkipAndTake_PR |        20 |        50 |     66.71 ns |   0.316 ns |   0.264 ns |     66.72 ns | 0.0422 |      - |     552 B |
|                     List_GetRange |        20 |        50 |     27.52 ns |   0.203 ns |   0.180 ns |     27.49 ns | 0.0349 | 0.0001 |     456 B |
|    ReadOnlyCollection_SkipAndTake |       200 |       500 |  1,948.57 ns |  12.762 ns |  11.313 ns |  1,946.40 ns | 0.3166 | 0.0038 |    4152 B |
| ReadOnlyCollection_SkipAndTake_PR |       200 |       500 |  1,894.11 ns |  10.677 ns |   9.987 ns |  1,891.91 ns | 0.3166 | 0.0038 |    4152 B |
|                  List_SkipAndTake |       200 |       500 |  1,360.73 ns |  11.972 ns |  10.612 ns |  1,362.17 ns | 0.3166 | 0.0038 |    4152 B |
|               List_SkipAndTake_PR |       200 |       500 |    218.02 ns |   4.362 ns |   6.919 ns |    215.36 ns | 0.3176 | 0.0045 |    4152 B |
|                     List_GetRange |       200 |       500 |    172.92 ns |   1.922 ns |   1.704 ns |    172.49 ns | 0.3102 | 0.0045 |    4056 B |
|    ReadOnlyCollection_SkipAndTake |      2000 |      5000 | 19,448.66 ns |  92.606 ns |  82.093 ns | 19,464.56 ns | 3.0518 | 0.3357 |   40152 B |
| ReadOnlyCollection_SkipAndTake_PR |      2000 |      5000 | 18,834.30 ns |  82.067 ns |  72.750 ns | 18,842.62 ns | 3.0518 | 0.3357 |   40152 B |
|                  List_SkipAndTake |      2000 |      5000 | 12,860.73 ns |  57.382 ns |  50.868 ns | 12,869.06 ns | 3.0670 | 0.3357 |   40152 B |
|               List_SkipAndTake_PR |      2000 |      5000 |  1,977.67 ns |  20.074 ns |  16.762 ns |  1,978.10 ns | 3.0670 | 0.3395 |   40152 B |
|                     List_GetRange |      2000 |      5000 |  1,957.41 ns |  13.694 ns |  12.139 ns |  1,955.80 ns | 3.0556 | 0.3395 |   40056 B |
```

```cs
[MemoryDiagnoser]
public class Benchmark
{
    [Params(2, 20, 200, 2000)]
    public int SkipCount { get; set; }

    [Params(5, 50, 500, 5000)]
    public int TakeCount { get; set; }

    private readonly List<string> _list = Enumerable.Range(0, 10000).Select(i => i.ToString()).ToList();
    private readonly ReadOnlyCollection<string> _readOnlyCollection = Enumerable.Range(0, 10000).Select(i => i.ToString()).ToList().AsReadOnly();

    [Benchmark]
    public List<string> ReadOnlyCollection_SkipAndTake() =>
        _readOnlyCollection.Skip(SkipCount).Take(TakeCount).ToList();

    [Benchmark]
    public List<string> ReadOnlyCollection_SkipAndTake_PR() =>
        _readOnlyCollection.Skip_PR(SkipCount).Take_PR(TakeCount).ToList();

    [Benchmark]
    public List<string> List_SkipAndTake() =>
        _list.Skip(SkipCount).Take(TakeCount).ToList();

    [Benchmark]
    public List<string> List_SkipAndTake_PR() =>
        _list.Skip_PR(SkipCount).Take_PR(TakeCount).ToList();

    [Benchmark]
    public List<string> List_GetRange() =>
        _list.GetRange(SkipCount, TakeCount);
}
```